Rename log.h to Logging.h
