+++ Common/include/ECUCANStructs.h
@@ -3,7 +3,7 @@
 
 #include "CANStruct.h"
 #include "dbc/structs/rivanna2.h"
-#include "log.h"
+#include "Logging.h"
 
 typedef struct ECUMotorCommands : CANStruct, rivanna2_ecu_motor_commands_t {
     void serialize(CANMessage *message) {
