+++ ECU/src/main.cpp
@@ -2,7 +2,7 @@
 #include "ECUCANStructs.h"
 #include "ECUInputReader.h"
 #include "Printing.h"
-#include "log.h"
+#include "Logging.h"
 #include "pindef.h"
 #include <mbed.h>
 #include <rtos.h>
