+++ Motor/src/main.cpp
@@ -3,7 +3,7 @@
 #include "MotorInterface.h"
 #include "MotorStateTracker.h"
 #include "Printing.h"
-#include "log.h"
+#include "Logging.h"
 #include "pindef.h"
 #include <mbed.h>
 #include <rtos.h>
