+++ Solar/src/main.cpp
@@ -1,7 +1,7 @@
 #include "Printing.h"
 #include "SolarCANInterface.h"
 #include "SolarInputReader.h"
-#include "log.h"
+#include "Logging.h"
 #include "mcp3008.h"
 #include "pindef.h"
 #include <mbed.h>
